notifier/pyproject.toml
[tool.poetry]
name = "aoc-bot"
version = "0.1.0"
description = ""
authors = ["Yuto Takano <moa17stock@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11, <3.14"
requests = "^2.32.5"
datargs = "^1.1.0"
apscheduler = "4.0.0a4"
hikari = {extras = ["server"], version = "^2.5.0"}
hikari-crescent = {extras = ["cron"], version = "^1.4.0"}

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
mypy = "^1.18.2"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
